\name{affy.pivot.table}
\alias{affy.pivot.table}
\title{merge expression and calls together}
\usage{
  affy.pivot.table(exp, calls, file = NULL)
}
\arguments{
  \item{exp}{a \code{matrix}-like object of expression
  levels}

  \item{calls}{a \code{matrix}-like object of detection
  calls (P/M/A)}

  \item{file}{the output filename}
}
\value{
  none.
}
\description{
  merge expression and calls together
}
\author{
  Mark Cowley
}

