\name{celDate}
\alias{celDate}
\title{datestamp of Affymetrix CEL file}
\usage{
  celDate(files)
}
\arguments{
  \item{files}{a vector of Affymertrix CEL files.}
}
\value{
  a vector of datestamps from the CEL file headers
}
\description{
  Extract the CEL file creation date stamp from within the
  CEL file header.
}
\author{
  Mark Cowley, 2008-07-29
}

