\name{hist_Pcount_cumulative}
\alias{hist_Pcount_cumulative}
\title{Plot a cumulative histogram of affy P/M/A calls.}
\usage{
  hist_Pcount_cumulative(Pcount = NULL, calls = NULL,
    grid = TRUE, plot = TRUE)
}
\arguments{
  \item{Pcount}{undocumented}

  \item{calls}{undocumented}

  \item{grid}{logical}

  \item{plot}{logical}
}
\value{
  none; generates a plot
}
\description{
  undocumented
}
\examples{
\dontrun{
  hist_Pcount_cumulative(calls=calls)
}
}
\author{
  Mark Cowley, 10 April 08
}

