\docType{package}
\name{mjcaffy-package}
\alias{mjcaffy}
\alias{mjcaffy-package}
\title{Helper functions for dealing with Affymetrix data.}
\description{
  \tabular{ll}{ Package: \tab mjcaffy\cr Type: \tab
  Package\cr Version: \tab 1.0.0\cr Date: \tab
  2012-07-17\cr License: \tab GPL-2\cr }
}
\author{
  Mark Cowley

  Maintainer: Mark Cowley <m.cowley@garvan.org.au>
}
\keyword{package}

