\name{test_parse.affy.ST.array.annotation}
\alias{test_parse.affy.ST.array.annotation}
\title{test_parse.affy.ST.array.annotation}
\usage{
  test_parse.affy.ST.array.annotation(verbose = TRUE)
}
\arguments{
  \item{verbose}{logical}
}
\value{
  nothing
}
\description{
  test_parse.affy.ST.array.annotation
}
\author{
  Mark Cowley, 2011-10-05
}

